navigation in containers and more
