+++ src/index.ts
@@ -3,4 +3,3 @@ export * from './Fragments';
 export * from './Inputs';
 export * from './Text';
 export * from './Utils';
-export * from './Navigation';
