+++ stories/Containers/Navigation.stories.tsx
@@ -4,7 +4,12 @@ import { AppleAlt } from '@styled-icons/fa-solid/AppleAlt';
 import { Archive } from '@styled-icons/fa-solid/Archive';
 import { BrowserRouter } from 'react-router-dom';
 import { Compass } from '@styled-icons/fa-solid/Compass';
-import { Navigation, NavigationProps, NavigationFooter } from '../../src';
+import {
+    Navigation,
+    NavigationProps,
+    NavigationFooter,
+    NavigationItem,
+} from '../../src';
 import { createStoryTitle } from '../Constants';
 import { logoWhite } from '../assets';
 
@@ -19,9 +24,45 @@ const defaultArgs = {
     label: 'Some Text', // this is for NavigationHeader
 };
 
+interface NavProps {
+    to: string;
+    label?: string;
+    icon: React.ForwardRefExoticComponent<React.RefAttributes<SVGSVGElement>>;
+}
+
+interface IPages {
+    [name: string]: {
+        component: React.LazyExoticComponent<React.ComponentType>;
+        navProps: NavProps;
+    };
+}
+
+const pages: IPages = {
+    Item1: {
+        component: null,
+        navProps: {
+            to: '/:id/orders',
+            icon: AppleAlt,
+        },
+    },
+    Item2: {
+        component: null,
+        navProps: {
+            to: '/:id/orders',
+            icon: Archive,
+        },
+    },
+};
+
 const Template: Story<NavigationProps> = (args) => (
     <BrowserRouter>
-        <Navigation {...args} />
+        <Navigation {...args}>
+            {Object.entries(pages).map(([key, { navProps }]) => (
+                <NavigationItem key={key} {...navProps}>
+                    {key}
+                </NavigationItem>
+            ))}
+        </Navigation>
     </BrowserRouter>
 );
 
@@ -29,26 +70,6 @@ export const Basic = Template.bind({});
 
 Basic.args = {
     ...defaultArgs,
-};
-
-export const WithPages = Template.bind({});
-
-WithPages.args = {
-    ...defaultArgs,
-    pages: {
-        Item1: {
-            navProps: {
-                to: '/:id/orders',
-                icon: AppleAlt,
-            },
-        },
-        Item2: {
-            navProps: {
-                to: '/:id/orders',
-                icon: Archive,
-            },
-        },
-    },
 };
 
 export const WithFooter = Template.bind({});
@@ -64,24 +85,10 @@ WithHeader.args = {
     header: <Header />,
 };
 
-export const WithPagesFooterHeader = Template.bind({});
-WithPagesFooterHeader.args = {
+export const WithFooterHeader = Template.bind({});
+WithFooterHeader.args = {
     ...defaultArgs,
     logo: logoWhite,
     footer: <NavigationFooter url="#foo" text="Link to" icon={Compass} />,
     header: <Header />,
-    pages: {
-        Item1: {
-            navProps: {
-                to: '/:id/orders',
-                icon: AppleAlt,
-            },
-        },
-        Item2: {
-            navProps: {
-                to: '/:id/orders',
-                icon: Archive,
-            },
-        },
-    },
 };
